moved tests to separate files
