+++ controller/controller/swarm_client_ctl.py
@@ -667,73 +667,3 @@ class SwarmClientController(ClientSideController):
             self.log_error(fl_ctx, f"Exception during _process_share_result: {secure_format_traceback()}")
             logger.error(f"Exception during _process_share_result: {e}")
             return make_reply(ReturnCode.EXECUTION_EXCEPTION)
-
-# Test Cases
-
-import unittest
-from unittest.mock import MagicMock
-
-class TestSwarmClientController(unittest.TestCase):
-
-    def setUp(self):
-        self.fl_ctx = MagicMock(FLContext)
-        self.controller = SwarmClientController(
-            task_name_prefix="test_prefix",
-            learn_task_name="test_learn_task",
-            persistor_id="test_persistor",
-            shareable_generator_id="test_generator",
-            aggregator_id="test_aggregator"
-        )
-
-    def test_initialization(self):
-        self.assertIsInstance(self.controller, SwarmClientController)
-        self.assertEqual(self.controller.task_name_prefix, "test_prefix")
-        self.assertEqual(self.controller.learn_task_name, "test_learn_task")
-
-    def test_process_config(self):
-        print("This test does not work yet.")
-        return
-        self.controller.me = "client1"
-        self.controller.get_config_prop = MagicMock(side_effect=lambda x, y=None: ["client1", "client2", "client3"])
-        self.controller.process_config(self.fl_ctx)
-        self.assertTrue(self.controller.is_trainer)
-        self.assertTrue(self.controller.is_aggr)
-
-    def test_execute(self):
-        print("This test does not work yet.")
-        return
-        shareable = MagicMock(Shareable)
-        abort_signal = MagicMock(Signal)
-        self.controller._process_learn_result = MagicMock(return_value=shareable)
-        result = self.controller.execute("test_learn_task", shareable, self.fl_ctx, abort_signal)
-        self.assertEqual(result, shareable)
-
-    def test_start_run(self):
-        print("This test does not work yet.")
-        return
-        self.controller.engine = MagicMock()
-        self.controller.engine.get_component = MagicMock(return_value=MagicMock(Aggregator))
-        self.controller.start_run(self.fl_ctx)
-        self.assertIsInstance(self.controller.aggregator, Aggregator)
-
-    def test_handle_event(self):
-        print("This test does not work yet.")
-        return
-        self.controller.me = "client1"
-        self.fl_ctx.get_prop = MagicMock(return_value="client1")
-        self.fl_ctx.set_prop = MagicMock()
-        self.controller.best_result = None
-        self.controller.handle_event(AppEventType.GLOBAL_BEST_MODEL_AVAILABLE, self.fl_ctx)
-        self.assertIsNotNone(self.controller.best_result)
-
-    def test_start_workflow(self):
-        print("This test does not work yet.")
-        return
-        shareable = MagicMock(Shareable)
-        abort_signal = MagicMock(Signal)
-        self.controller._scatter = MagicMock(return_value=True)
-        result = self.controller.start_workflow(shareable, self.fl_ctx, abort_signal)
-        self.assertEqual(result.get_return_code(), ReturnCode.OK)
-
-if __name__ == "__main__":
-    unittest.main()
