+++ controller/controller/swarm_server_ctl.py
@@ -96,56 +96,3 @@ class SwarmServerController(ServerSideController):
         except Exception as e:
             logger.error(f"Error during prepare_config: {e}")
             raise
-
-# Test Cases
-
-import unittest
-from unittest.mock import MagicMock
-
-class TestSwarmServerController(unittest.TestCase):
-
-    def setUp(self):
-        self.fl_ctx = MagicMock(FLContext)
-        self.participating_clients = ['client1', 'client2', 'client3']
-        self.controller = SwarmServerController(
-            num_rounds=10,
-            participating_clients=self.participating_clients,
-            starting_client='client1'
-        )
-
-    def test_initialization(self):
-        self.assertIsInstance(self.controller, SwarmServerController)
-        self.assertEqual(self.controller.num_rounds, 10)
-        self.assertEqual(self.controller.starting_client, 'client1')
-
-    def test_start_controller(self):
-        print("This test does not work yet.")
-        return
-        self.controller.start_controller(self.fl_ctx)
-        self.assertIn('client1', self.controller.train_clients)
-        self.assertIn('client1', self.controller.aggr_clients)
-
-    def test_prepare_config(self):
-        config = self.controller.prepare_config()
-        self.assertIn(Constant.AGGR_CLIENTS, config)
-        self.assertIn(Constant.TRAIN_CLIENTS, config)
-
-    def test_invalid_starting_client(self):
-        print("This test does not work yet.")
-        return
-        with self.assertRaises(ValueError):
-            SwarmServerController(
-                num_rounds=10,
-                participating_clients=self.participating_clients,
-                starting_client=None
-            )
-
-    def test_client_not_in_train_or_aggr(self):
-        print("This test does not work yet.")
-        return
-        self.controller.participating_clients.append('client4')
-        with self.assertRaises(RuntimeError):
-            self.controller.start_controller(self.fl_ctx)
-
-if __name__ == "__main__":
-    unittest.main()
